cleaning

+++ TERM_1/hashTableLibrary.h
@@ -26,11 +26,11 @@ void deleteElementFromTable(struct HashTable* table, char* key);
 struct Node* findElement(struct HashTable* table, char* key);
 int getValue(struct HashTable* table, char* key);
 void set(struct HashTable* table, char* key, int value);
-int numberOfChains(struct HashTable* table);
-int numberOfElemnts(struct HashTable* table);
-int minChainLength(struct HashTable* table);
-int maxChainLength(struct HashTable* table);
-int averageChainLength(struct HashTable* table);
+size_t numberOfChains(struct HashTable* table);
+size_t numberOfElemnts(struct HashTable* table);
+size_t minChainLength(struct HashTable* table);
+size_t maxChainLength(struct HashTable* table);
+size_t averageChainLength(struct HashTable* table);
 void getInformation(struct HashTable* myHashTable, struct Information* myInf);
 
 #endif HASHTABLE_LIBRARY_INCLUDED
